beforeCarlosDestroy

+++ cadicom/src/main/java/es/uca/cadicom/repository/FacturaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface FacturaRepository extends JpaRepository<Factura, Integer> {
-    List<Factura> findByTelefonoId(Long telefono_id);
+    List<Factura> findByTelefonoId(Integer telefono_id);
 }
